refactor(reservation-app): seat 기능을 book 패키지에서 독립된 기능 패키지로 분리함
