--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/application/port/in/CreateSeatUseCase.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.application.port.in;
-
-import com.seatliberator.seatliberator.reservation.book.application.port.in.command.CreateSeatCommand;
-
-public interface CreateSeatUseCase {
-    boolean create(CreateSeatCommand command);
-}
