--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/application/port/in/DeleteSeatUseCase.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.application.port.in;
-
-import com.seatliberator.seatliberator.reservation.book.application.port.in.command.DeleteSeatCommand;
-
-public interface DeleteSeatUseCase {
-    boolean delete(DeleteSeatCommand command);
-}
