--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/application/port/in/UpdateSeatUseCase.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.application.port.in;
-
-import com.seatliberator.seatliberator.reservation.book.application.port.in.command.UpdateSeatCommand;
-
-public interface UpdateSeatUseCase {
-    boolean update(UpdateSeatCommand command);
-}
