--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/application/port/in/command/CreateSeatCommand.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.application.port.in.command;
-
-public record CreateSeatCommand(
-        String roomId,
-        String seatId
-) {
-}
