--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/application/port/in/command/DeleteSeatCommand.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.application.port.in.command;
-
-public record DeleteSeatCommand(
-        String roomId,
-        String seatId
-) {
-}
