+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/seat/application/port/in/DeleteSeatUseCase.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.reservation.seat.application.port.in;
+
+import com.seatliberator.seatliberator.reservation.seat.application.port.in.command.DeleteSeatCommand;
+
+public interface DeleteSeatUseCase {
+    boolean delete(DeleteSeatCommand command);
+}
