+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/seat/application/port/in/UpdateSeatUseCase.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.reservation.seat.application.port.in;
+
+import com.seatliberator.seatliberator.reservation.seat.application.port.in.command.UpdateSeatCommand;
+
+public interface UpdateSeatUseCase {
+    boolean update(UpdateSeatCommand command);
+}
