+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/seat/application/service/SeatQueryService.java
@@ -1,9 +1,9 @@
-package com.seatliberator.seatliberator.reservation.book.application.service;
+package com.seatliberator.seatliberator.reservation.seat.application.service;
 
-import com.seatliberator.seatliberator.reservation.book.application.port.in.FindSeatUseCase;
-import com.seatliberator.seatliberator.reservation.book.application.port.in.result.SeatResult;
-import com.seatliberator.seatliberator.reservation.book.application.port.out.SeatReader;
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
+import com.seatliberator.seatliberator.reservation.seat.application.port.in.FindSeatUseCase;
+import com.seatliberator.seatliberator.reservation.seat.application.port.in.result.SeatResult;
+import com.seatliberator.seatliberator.reservation.seat.application.port.out.SeatReader;
 import com.seatliberator.seatliberator.reservation.shared.application.exception.ReservationApplicationErrorCode;
 import com.seatliberator.seatliberator.reservation.shared.application.exception.ReservationApplicationException;
 import lombok.RequiredArgsConstructor;
