+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/seat/infrastructure/web/request/SeatUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.book.infrastructure.web.request;
+package com.seatliberator.seatliberator.reservation.seat.infrastructure.web.request;
 
 public record SeatUpdateRequest(
         String oldRoomId,
