+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/integration/VacancyRequestTest.java
@@ -3,9 +3,9 @@ package com.seatliberator.seatliberator.reservation.integration;
 import com.seatliberator.seatliberator.reservation.VacancyAlertRequestCreateCommandBuilder;
 import com.seatliberator.seatliberator.reservation.book.application.port.in.CreateReservationUseCase;
 import com.seatliberator.seatliberator.reservation.book.application.port.in.command.CreateReservationCommand;
-import com.seatliberator.seatliberator.reservation.book.application.port.in.command.CreateSeatCommand;
-import com.seatliberator.seatliberator.reservation.book.application.service.SeatCommandService;
 import com.seatliberator.seatliberator.reservation.domain.VacancyAlertRequestStatus;
+import com.seatliberator.seatliberator.reservation.seat.application.port.in.command.CreateSeatCommand;
+import com.seatliberator.seatliberator.reservation.seat.application.service.SeatCommandService;
 import com.seatliberator.seatliberator.reservation.shared.application.exception.ReservationApplicationException;
 import com.seatliberator.seatliberator.reservation.vacancy.application.port.in.RequestVacancyAlertUseCase;
 import com.seatliberator.seatliberator.reservation.vacancy.application.port.out.VacancyAlertRequestStore;
